# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "CloudServiceStore.WebApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Auth.Commands.Register.RegisterCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Auth.Commands.Register.RegisterCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Auth.Commands.Register.RegisterCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Auth.Commands.Login.LoginCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Auth.Commands.Login.LoginCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Auth.Commands.Login.LoginCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/refresh-token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.WebApi.Controllers.RefreshTokenRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.WebApi.Controllers.RefreshTokenRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.WebApi.Controllers.RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.WebApi.Controllers.ForgotPasswordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.WebApi.Controllers.ForgotPasswordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.WebApi.Controllers.ForgotPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Auth.Commands.ResetPassword.ResetPasswordCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Auth.Commands.ResetPassword.ResetPasswordCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Auth.Commands.ResetPassword.ResetPasswordCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/google-login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Auth.Commands.GoogleLogin.GoogleLoginCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Auth.Commands.GoogleLogin.GoogleLoginCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Auth.Commands.GoogleLogin.GoogleLoginCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/auth/2fa/enable": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.WebApi.Controllers.EnableTwoFactorRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.WebApi.Controllers.EnableTwoFactorRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.WebApi.Controllers.EnableTwoFactorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/2fa/verify-login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.WebApi.Controllers.VerifyTwoFactorRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.WebApi.Controllers.VerifyTwoFactorRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.WebApi.Controllers.VerifyTwoFactorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/banners": {
      "get": {
        "tags": [
          "Banners"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Banners"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Banners.Commands.CreateBanner.CreateBannerCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Banners.Commands.CreateBanner.CreateBannerCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Banners.Commands.CreateBanner.CreateBannerCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/banners/{id}": {
      "put": {
        "tags": [
          "Banners"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Banners.Commands.UpdateBanner.UpdateBannerCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Banners.Commands.UpdateBanner.UpdateBannerCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloudServiceStore.Application.Features.Banners.Commands.UpdateBanner.UpdateBannerCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Banners"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/banners/upload": {
      "post": {
        "tags": [
          "Banners"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/banners/{id}/image": {
      "post": {
        "tags": [
          "Banners"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CloudServiceStore.Application.Features.Auth.Commands.GoogleLogin.GoogleLoginCommand": {
        "type": "object",
        "properties": {
          "credential": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CloudServiceStore.Application.Features.Auth.Commands.Login.LoginCommand": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "ipAddress": {
            "type": "string",
            "nullable": true
          },
          "userAgent": {
            "type": "string",
            "nullable": true
          },
          "deviceInfo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CloudServiceStore.Application.Features.Auth.Commands.Register.RegisterCommand": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "ward": {
            "type": "string",
            "nullable": true
          },
          "addressLine": {
            "type": "string",
            "nullable": true
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "taxCode": {
            "type": "string",
            "nullable": true
          },
          "subscribeNewsletter": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CloudServiceStore.Application.Features.Auth.Commands.ResetPassword.ResetPasswordCommand": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "nullable": true
          },
          "newPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CloudServiceStore.Application.Features.Banners.Commands.CreateBanner.CreateBannerCommand": {
        "type": "object",
        "properties": {
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "linkUrl": {
            "type": "string",
            "nullable": true
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CloudServiceStore.Application.Features.Banners.Commands.UpdateBanner.UpdateBannerCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "imageUrl": {
            "type": "string",
            "nullable": true
          },
          "linkUrl": {
            "type": "string",
            "nullable": true
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CloudServiceStore.WebApi.Controllers.EnableTwoFactorRequest": {
        "type": "object",
        "properties": {
          "secretKey": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CloudServiceStore.WebApi.Controllers.ForgotPasswordRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CloudServiceStore.WebApi.Controllers.RefreshTokenRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CloudServiceStore.WebApi.Controllers.VerifyTwoFactorRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}